Issue #108
When the current tab is the preferences tab, and you call save all,
nothing happens. Also it is possible, that before this the save all just
saved the current tab one time for each opened tab.

Pending: When the save as dialog opens because of save all, change
current tab to the tab being saved.